test: mapped interest rates related tests
